feat: Separated the File and Edit Some Motion Feature #31
