+++ src/App.tsx
@@ -9,7 +9,7 @@ import { InfoHowLayout } from './layout/info/InfoHowLayout'
 import { InfowwwLayout } from './layout/info/InfowwwLayout'
 import { InfoCreditLayout } from './layout/info/InfoCreditLayout'
 import { InfoFontLayout } from './layout/info/InfoFontLayout'
-import { ComingSoon } from './layout/common/ComingSoon'
+import { ComingSoon } from './components/common/ComingSoon'
 import { isMobile } from 'react-device-detect'
 
 const App = () => {
@@ -24,21 +24,21 @@ const App = () => {
   }, 3990)
 
   return (
-    // <ComingSoon />
-    <>
-      <GlobalHeader />
-      <Routes>
-        <Route path="/" element={<HomePage loading={loading} render={render} />} />
-        <Route path="/www" element={<InfoPage />}>
-          <Route path="" element={<InfowwwLayout />} />
-          <Route path="how" element={<InfoHowLayout />} />
-          <Route path="credit" element={<InfoCreditLayout />} />
-          <Route path="font" element={<InfoFontLayout />} />
-        </Route>
-        <Route path="/artwork" element={<ArtworkPage />} />
-        <Route path="/artist" element={<ArtistPage />} />
-      </Routes>
-    </>
+    <ComingSoon />
+    // <>
+    //   <GlobalHeader />
+    //   <Routes>
+    //     <Route path="/" element={<HomePage loading={loading} render={render} />} />
+    //     <Route path="/www" element={<InfoPage />}>
+    //       <Route path="" element={<InfowwwLayout />} />
+    //       <Route path="how" element={<InfoHowLayout />} />
+    //       <Route path="credit" element={<InfoCreditLayout />} />
+    //       <Route path="font" element={<InfoFontLayout />} />
+    //     </Route>
+    //     <Route path="/artwork" element={<ArtworkPage />} />
+    //     <Route path="/artist" element={<ArtistPage />} />
+    //   </Routes>
+    // </>
   )
 }
